pyproject.toml
[project]
name = "silverforge"
version = "0.1.0"
description = "PDF를 구조화된 Markdown으로 변환 - Upstage Document Parse 기반"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Harrison Kim"}
]
keywords = ["pdf", "markdown", "document-parsing", "upstage", "llm", "data-pipeline"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "langchain-upstage>=0.1.0",
    "jiwer>=3.0.0",
    # UI Console
    "streamlit>=1.30.0",
    "watchdog>=4.0.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
silverforge-ui = "silverforge.app:main"
silverforge-api = "silverforge.api:run_server"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/silverforge"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/scripts",
    "/README.md",
    "/README.ko.md",
]

[tool.ruff]
line-length = 100
target-version = "py310"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
